requirements.txt
torch
numpy>=1.24.0
Flask>=3.0.0
pydub>=0.25.0
soundfile>=0.12.0
sounddevice>=0.4.6
openai-whisper>=20231117
langchain-ollama>=0.2.0
langchain-chroma>=0.1.0
langchain-community>=0.2.0
requests>=2.31.0
python-docx>=1.1.0
docx2txt>=0.8